config/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.1
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn>=21.0.0
